se double instead of float to preserve original JSON data

+++ jgltf-impl-v1/src/main/java/de/javagl/jgltf/impl/v1/TechniqueStatesFunctions.java
@@ -30,7 +30,7 @@ public class TechniqueStatesFunctions
      * &nbsp;&nbsp;The elements of this array (optional) 
      * 
      */
-    private float[] blendColor;
+    private double[] blendColor;
     /**
      * Integer values passed to `blendEquationSeparate()`. (optional)<br> 
      * Default: [32774,32774]<br> 
@@ -100,7 +100,7 @@ public class TechniqueStatesFunctions
      * &nbsp;&nbsp;The elements of this array (optional) 
      * 
      */
-    private float[] depthRange;
+    private double[] depthRange;
     /**
      * Integer value passed to `frontFace()`. (optional)<br> 
      * Default: [2305]<br> 
@@ -120,7 +120,7 @@ public class TechniqueStatesFunctions
      * &nbsp;&nbsp;Minimum: 0.0 (exclusive) 
      * 
      */
-    private float[] lineWidth;
+    private double[] lineWidth;
     /**
      * Floating-point value passed to `polygonOffset()`. [factor, units] 
      * (optional)<br> 
@@ -130,7 +130,7 @@ public class TechniqueStatesFunctions
      * &nbsp;&nbsp;The elements of this array (optional) 
      * 
      */
-    private float[] polygonOffset;
+    private double[] polygonOffset;
     /**
      * Floating-point value passed to `scissor()`. [x, y, width, height]. 
      * (optional)<br> 
@@ -140,7 +140,7 @@ public class TechniqueStatesFunctions
      * &nbsp;&nbsp;The elements of this array (optional) 
      * 
      */
-    private float[] scissor;
+    private double[] scissor;
 
     /**
      * Floating-point values passed to `blendColor()`. [red, green, blue, 
@@ -155,7 +155,7 @@ public class TechniqueStatesFunctions
      * the given constraints
      * 
      */
-    public void setBlendColor(float[] blendColor) {
+    public void setBlendColor(double[] blendColor) {
         if (blendColor == null) {
             this.blendColor = blendColor;
             return ;
@@ -180,7 +180,7 @@ public class TechniqueStatesFunctions
      * @return The blendColor
      * 
      */
-    public float[] getBlendColor() {
+    public double[] getBlendColor() {
         return this.blendColor;
     }
 
@@ -191,8 +191,8 @@ public class TechniqueStatesFunctions
      * @return The default blendColor
      * 
      */
-    public float[] defaultBlendColor() {
-        return new float[] { 0.0F, 0.0F, 0.0F, 0.0F };
+    public double[] defaultBlendColor() {
+        return new double[] { 0.0F, 0.0F, 0.0F, 0.0F };
     }
 
     /**
@@ -546,7 +546,7 @@ public class TechniqueStatesFunctions
      * the given constraints
      * 
      */
-    public void setDepthRange(float[] depthRange) {
+    public void setDepthRange(double[] depthRange) {
         if (depthRange == null) {
             this.depthRange = depthRange;
             return ;
@@ -571,7 +571,7 @@ public class TechniqueStatesFunctions
      * @return The depthRange
      * 
      */
-    public float[] getDepthRange() {
+    public double[] getDepthRange() {
         return this.depthRange;
     }
 
@@ -582,8 +582,8 @@ public class TechniqueStatesFunctions
      * @return The default depthRange
      * 
      */
-    public float[] defaultDepthRange() {
-        return new float[] { 0.0F, 1.0F };
+    public double[] defaultDepthRange() {
+        return new double[] { 0.0F, 1.0F };
     }
 
     /**
@@ -657,7 +657,7 @@ public class TechniqueStatesFunctions
      * the given constraints
      * 
      */
-    public void setLineWidth(float[] lineWidth) {
+    public void setLineWidth(double[] lineWidth) {
         if (lineWidth == null) {
             this.lineWidth = lineWidth;
             return ;
@@ -668,7 +668,7 @@ public class TechniqueStatesFunctions
         if (lineWidth.length > 1) {
             throw new IllegalArgumentException("Number of lineWidth elements is > 1");
         }
-        for (float lineWidthElement: lineWidth) {
+        for (double lineWidthElement: lineWidth) {
             if (lineWidthElement<= 0.0D) {
                 throw new IllegalArgumentException("lineWidthElement <= 0.0");
             }
@@ -687,7 +687,7 @@ public class TechniqueStatesFunctions
      * @return The lineWidth
      * 
      */
-    public float[] getLineWidth() {
+    public double[] getLineWidth() {
         return this.lineWidth;
     }
 
@@ -698,8 +698,8 @@ public class TechniqueStatesFunctions
      * @return The default lineWidth
      * 
      */
-    public float[] defaultLineWidth() {
-        return new float[] { 1.0F };
+    public double[] defaultLineWidth() {
+        return new double[] { 1.0F };
     }
 
     /**
@@ -715,7 +715,7 @@ public class TechniqueStatesFunctions
      * the given constraints
      * 
      */
-    public void setPolygonOffset(float[] polygonOffset) {
+    public void setPolygonOffset(double[] polygonOffset) {
         if (polygonOffset == null) {
             this.polygonOffset = polygonOffset;
             return ;
@@ -740,7 +740,7 @@ public class TechniqueStatesFunctions
      * @return The polygonOffset
      * 
      */
-    public float[] getPolygonOffset() {
+    public double[] getPolygonOffset() {
         return this.polygonOffset;
     }
 
@@ -751,8 +751,8 @@ public class TechniqueStatesFunctions
      * @return The default polygonOffset
      * 
      */
-    public float[] defaultPolygonOffset() {
-        return new float[] { 0.0F, 0.0F };
+    public double[] defaultPolygonOffset() {
+        return new double[] { 0.0F, 0.0F };
     }
 
     /**
@@ -768,7 +768,7 @@ public class TechniqueStatesFunctions
      * the given constraints
      * 
      */
-    public void setScissor(float[] scissor) {
+    public void setScissor(double[] scissor) {
         if (scissor == null) {
             this.scissor = scissor;
             return ;
@@ -793,7 +793,7 @@ public class TechniqueStatesFunctions
      * @return The scissor
      * 
      */
-    public float[] getScissor() {
+    public double[] getScissor() {
         return this.scissor;
     }
 
@@ -804,8 +804,8 @@ public class TechniqueStatesFunctions
      * @return The default scissor
      * 
      */
-    public float[] defaultScissor() {
-        return new float[] { 0.0F, 0.0F, 0.0F, 0.0F };
+    public double[] defaultScissor() {
+        return new double[] { 0.0F, 0.0F, 0.0F, 0.0F };
     }
 
 }
